refactor routes
